Add Check Data step to AddTable
